Some refactoring, moved helper functions to helpers.h
